moved fetch account code
